Continue implementation of timelapse camera feature.

+++ src/main/java/celtech/roboxbase/configuration/datafileaccessors/CameraProfileContainer.java
@@ -1,6 +1,8 @@
 package celtech.roboxbase.configuration.datafileaccessors;
 
+import celtech.roboxbase.comms.DetectedServer;
 import celtech.roboxbase.configuration.BaseConfiguration;
+import celtech.roboxbase.configuration.CoreMemory;
 import celtech.roboxbase.configuration.fileRepresentation.CameraProfile;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
@@ -26,12 +28,14 @@ public class CameraProfileContainer
     private static CameraProfileContainer instance;
     
     private static Map<String, CameraProfile> cameraProfilesMap;
-    
+ 
+    private static String anyCameraName = "";
+
     private CameraProfileContainer()
     {
         cameraProfilesMap = new HashMap<>();
         CameraProfile defaultCameraProfile = new CameraProfile();
-        cameraProfilesMap.put(defaultCameraProfile.getProfileName(), defaultCameraProfile);
+        cameraProfilesMap.put(defaultCameraProfile.getProfileName().toLowerCase(), defaultCameraProfile);
         loadCameraProfiles();
     }
     
@@ -58,7 +62,7 @@ public class CameraProfileContainer
             for(Path path : stream)
             {
                 CameraProfile cameraProfile = objectMapper.readValue(path.toFile(), CameraProfile.class);
-                cameraProfilesMap.put(cameraProfile.getProfileName(), cameraProfile);
+                cameraProfilesMap.put(cameraProfile.getProfileName().toLowerCase(), cameraProfile);
             }
         } catch (IOException ex) 
         {
@@ -68,49 +72,74 @@ public class CameraProfileContainer
     
     public void saveCameraProfile(CameraProfile cameraProfile)
     {
-        ObjectMapper objectMapper = new ObjectMapper();
-        objectMapper.registerModule(new Jdk8Module());
-        
-        String cameraProfileFileName = cameraProfile.getProfileName() + BaseConfiguration.cameraProfileFileExtention;
-        Path cameraProfilePath = Path.of(BaseConfiguration.getUserCameraProfileDirectory() + File.separator + cameraProfileFileName);
+        if (cameraProfile.getProfileName().equals(BaseConfiguration.defaultCameraProfileName))
+        {
+            STENO.warning("Can't save default camera profile.");
+        }
+        else
+        {
+            ObjectMapper objectMapper = new ObjectMapper();
+            objectMapper.registerModule(new Jdk8Module());
 
-        try 
-        {
-            objectMapper.writerWithDefaultPrettyPrinter().writeValue(cameraProfilePath.toFile(), cameraProfile);
-            cameraProfilesMap.put(cameraProfile.getProfileName(), cameraProfile);
-        } catch (IOException ex) 
-        {
-            STENO.exception("Error when trying to save profile of " + cameraProfilePath.toString(), ex);
+            String cameraProfileFileName = cameraProfile.getProfileName() + BaseConfiguration.cameraProfileFileExtention;
+            Path cameraProfilePath = Path.of(BaseConfiguration.getUserCameraProfileDirectory() + File.separator + cameraProfileFileName);
+
+            try 
+            {
+                objectMapper.writerWithDefaultPrettyPrinter().writeValue(cameraProfilePath.toFile(), cameraProfile);
+                cameraProfilesMap.put(cameraProfile.getProfileName().toLowerCase(), cameraProfile);
+            } catch (IOException ex) 
+            {
+                STENO.exception("Error when trying to save profile of " + cameraProfilePath.toString(), ex);
+            }
         }
     }
     
     public void deleteCameraProfile(CameraProfile cameraProfile)
     {
-        String profileName = cameraProfile.getProfileName();
-        
-        if (cameraProfilesMap.containsKey(profileName))
+        if (cameraProfile.getProfileName().equals(BaseConfiguration.defaultCameraProfileName))
         {
-            cameraProfilesMap.remove(profileName);
-        } else
-        {
-            STENO.error("File " + profileName + ", doesn't exist");
+            STENO.warning("Can't delete default camera profile.");
         }
-        
-        String filePath = BaseConfiguration.getUserCameraProfileDirectory() + File.separator 
-                   + profileName + BaseConfiguration.cameraProfileFileExtention;
-        Path fileToDelete = Paths.get(filePath);
-        
-        try 
+        else
         {
-            Files.deleteIfExists(fileToDelete);
-        } catch (IOException ex) 
-        {
-            STENO.exception("Error when trying to delete profile of " + filePath, ex);
+            String profileName = cameraProfile.getProfileName();
+            String profileNameLC = profileName.toLowerCase();
+
+            if (cameraProfilesMap.containsKey(profileNameLC))
+            {
+                cameraProfilesMap.remove(profileNameLC);
+            } else
+            {
+                STENO.error("File " + profileName + ", doesn't exist");
+            }
+
+            String filePath = BaseConfiguration.getUserCameraProfileDirectory() + File.separator 
+                       + profileName + BaseConfiguration.cameraProfileFileExtention;
+            Path fileToDelete = Paths.get(filePath);
+
+            try 
+            {
+                Files.deleteIfExists(fileToDelete);
+            } catch (IOException ex) 
+            {
+                STENO.exception("Error when trying to delete profile of " + filePath, ex);
+            }
         }
     }
     
     public Map<String, CameraProfile> getCameraProfilesMap()
     {
         return cameraProfilesMap;
+    }
+
+    public CameraProfile getProfileByName(String profileName)
+    {
+        return cameraProfilesMap.get(profileName.toLowerCase());
+    }
+
+    public CameraProfile getDefaultProfile()
+    {
+        return cameraProfilesMap.get(BaseConfiguration.defaultCameraProfileName.toLowerCase());
     }
 }
